Finished final worksheet, added comments
Finished Multiple Conditions project Tire Pressure 1. Added an array
with 4 places, allowing a user prompt to enter the values of each
placeholder. If statement checks to see if all the values match where
they should and executes output appropriately to the boolean value.

Also added proper header in comments for each section of the worksheet.